Replace context stack instead of clearing it
This commit changes the `Context.clear` method to replace the context
stack with a new empty array instead of mutating the existing array.

Rails copies all thread locals to a new thread for certain requests [1],
which means the stack may be shared between threads. In that case,
modifying the array in one thread would affect the other thread.

[1]: https://github.com/rails/rails/blob/3163bb735189eddd7c152a1e72bed7b4d052f7d4/actionpack/lib/action_controller/metal/live.rb#L285-L288